+++ .internal/assemblyscript/@gl/api/w2h/player.ts
@@ -1,6 +1,5 @@
 import { PlayerAction } from "../../utils/movement/topdown";
 
-export declare function setVariation(variation: string): void;
 export declare function setPos(x: f32, y: f32): void;
 export declare function setSize(width: f32, height: f32): void;
 export declare function setPivot(x: f32, y: f32): void;
@@ -14,8 +13,8 @@ export declare function setCollisionBox(
   width: f32,
   height: f32
 ): void;
+export declare function setSkin(skin: string): void;
 
-export const _keep_setVariation = setVariation;
 export const _keep_setPos = setPos;
 // export const _keep_setSize = setSize;
 export const _keep_setPivot = setPivot;
@@ -23,4 +22,6 @@ export const _keep_setZIndex = setZIndex;
 export const _keep_setAction = setAction;
 export const _keep_flip = flip;
 export const _keep_setSpeed = setSpeed;
+export const _keep_setSkin = setSkin;
+
 // export const _keep_setCollisionBox = setCollisionBox;